+++ src/cow-react/modules/wallet/web3-react/containers/AddToMetamask/index.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react'
 import { Currency } from '@uniswap/sdk-core'
 
 import { useWeb3React } from '@web3-react/core'
-import { getIsMetaMask } from '@cow/modules/wallet/web3-react/utils/connection'
+import { getIsMetaMask } from '@cow/modules/wallet/api/utils/connections'
 import { addTokenToMetamaskAnalytics } from 'components/analytics'
 import useCurrencyLogoURIs from 'lib/hooks/useCurrencyLogoURIs'
 import { AddToMetamask as AddToMetamaskPure } from '../../../api/pure/AddToMetamask'
