+++ src/custom/utils/switchChain.ts
@@ -6,7 +6,7 @@ import {
   injectedConnection,
   networkConnection,
   walletConnectConnection,
-} from '@cow/modules/wallet/web3-react/utils/connection/connections'
+} from '@cow/modules/wallet/web3-react/utils/connection'
 import { getChainInfo } from 'constants/chainInfo'
 import { ALL_SUPPORTED_CHAIN_IDS, SupportedChainId } from 'constants/chains'
 import { RPC_URLS } from 'constants/networks'
